refactor(agentic): consolidate deriveWorktreePath + pane lookups
Eliminate three copies of deriveWorktreePath (spawn-helper, agentic-core
MCP tool, agentic extension) and two copies of paneExists/lookupPaneIntendedAgent
(agentic-core MCP tool, agent-detection-service). The duplicates had
already drifted: spawn-helper used Unix-only regex while the others
accepted both separators.

- Export deriveWorktreePath from spawn-helper with cross-platform regex.
- Add src/lib/services/pane-lookup.ts as the single home for paneExists
  + lookupPaneIntendedAgent; importers updated.
- Surface deriveWorktreePath on ExtensionAPI so the agentic extension
  stays within its barrier instead of re-implementing.
