+++ src/lib/services/extension-api.ts
@@ -88,6 +88,7 @@ import {
   pushExternalAttention as corePushExternalAttention,
 } from "./attention-api";
 import { agentsStore } from "./agent-detection-service";
+import { deriveWorktreePath as coreDeriveWorktreePath } from "./spawn-helper";
 import { waitRestored } from "../bootstrap/restore-workspaces";
 import type { WorkspaceTemplate } from "../config";
 import type { WorkspaceDefInput } from "../../extensions/api";
@@ -369,6 +370,9 @@ export function createExtensionAPI(
       const all = get(agentsStore);
       const hit = all.find((a) => a.paneId === paneId);
       return (hit ?? null) as ReturnType<ExtensionAPI["getAgentByPane"]>;
+    },
+    deriveWorktreePath(repoPath: string, branch: string): string {
+      return coreDeriveWorktreePath(repoPath, branch);
     },
     getAllTerminalSurfaces() {
       const out: Array<{ id: string; workspaceId: string; title: string }> = [];
